+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [2.4.0] - 2017-12-18
+
+### Added
+- Nakadi collects access log events
+- Sending of KPI metrics for event-types count
+
 ## [2.3.4] - 2017-12-14
 
 ### Changed